Fix memory leak in compression level with non-Integer
This PR fixes a memory leak when an invalid level option is passed to Zstd.compress.

If non-Integer object was given as compression level,
NUM2INT cause a TypeError exception and memory leak.

This changes will release the context if necessary and raise an exception
if non-Integer object was given as compression level.